refactor(client): split ws_client/client.rs into focused modules
Split the 456-line client.rs grab-bag so each file owns one
responsibility and the session state vocabulary lives with the state
machine that produces it (#475/#476):

- session/types.rs: WsClientError / WsClientResult / InvalidStateKind
  / WsCloseReason. session.rs no longer reverse-imports its own state
  types from client.rs — it owns them via `pub use types::{...}`.
- dispatcher.rs: EventDispatcherHandler + do_without_validation, plus
  its sole consumers EventHandler trait and RawEventEnvelope/Header.
- client.rs: only LarkWsClient + endpoint discovery (open/open_with/
  get_conn_url) + EndPointResponse/ClientConfig serde types.

The public `pub use` surface in ws_client/mod.rs is unchanged — all
seven names still resolve at ws_client::*, only their source module
moved. Non-breaking; frame_handler.rs gets the one inevitable import
split, package.rs/headers.rs untouched.

ws_client behavior unchanged: 174 existing tests green, clippy
-Dwarnings clean, cargo doc clean, websocket_echo_bot example builds.

Closes #476
