+++ crates/openlark-client/src/ws_client/client.rs
@@ -9,14 +9,11 @@ use log::{debug, info};
 use reqwest::Client;
 use serde::Deserialize;
 use serde_json::json;
-use tokio::sync::mpsc;
 use tokio_tungstenite::connect_async_with_config;
-use tokio_tungstenite::tungstenite::protocol::frame::coding::CloseCode;
 use url::Url;
 
-use super::session::{Session, SessionOptions};
-
-type EventHandlerResult = Result<(), Box<dyn std::error::Error + Send + Sync>>;
+use super::dispatcher::EventDispatcherHandler;
+use super::session::{Session, SessionOptions, WsClientError, WsClientResult};
 
 /// WebSocket endpoint API 专用响应结构（顶层 code/msg/data）
 #[derive(Debug, Deserialize)]
@@ -51,133 +48,6 @@ fn extract_endpoint_response(
     }
 
     Ok(end_point)
-}
-
-#[derive(Debug, Deserialize)]
-struct RawEventEnvelope {
-    header: RawEventHeader,
-}
-
-#[derive(Debug, Deserialize)]
-struct RawEventHeader {
-    #[serde(default)]
-    event_type: String,
-}
-
-/// 原始事件处理器。
-///
-/// 当调用方希望直接消费 WebSocket 原始事件负载时，可以实现该 trait，
-/// 再通过 [`EventDispatcherHandler::register_raw`] 注册：
-///
-/// - key=`"raw"`：接收所有原始事件负载
-/// - key=`"<event_type>"`：仅接收指定 `header.event_type` 的事件
-pub trait EventHandler: Send + Sync + 'static {
-    /// 处理原始事件负载。
-    fn handle(&self, payload: &[u8]) -> EventHandlerResult;
-}
-
-/// WebSocket 事件分发处理器。
-///
-/// 目前支持两类分发目标：
-///
-/// - `payload_sender(...)`：把原始负载转发到 channel
-/// - `register_raw(...)`：注册原始事件处理器
-#[derive(Clone)]
-pub struct EventDispatcherHandler {
-    payload_tx: Option<mpsc::UnboundedSender<Vec<u8>>>,
-    raw_handlers: HashMap<String, Arc<dyn EventHandler>>,
-}
-
-impl std::fmt::Debug for EventDispatcherHandler {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("EventDispatcherHandler")
-            .field(
-                "payload_tx",
-                &self.payload_tx.as_ref().map(|_| "configured"),
-            )
-            .field(
-                "raw_handler_keys",
-                &self.raw_handlers.keys().collect::<Vec<_>>(),
-            )
-            .finish()
-    }
-}
-
-impl EventDispatcherHandler {
-    /// 通配原始事件处理器 key。
-    pub const RAW_EVENT_KEY: &'static str = "raw";
-
-    /// 创建新的事件分发构建器。
-    pub fn builder() -> Self {
-        Self {
-            payload_tx: None,
-            raw_handlers: HashMap::new(),
-        }
-    }
-
-    /// 完成构建。
-    pub fn build(self) -> Self {
-        self
-    }
-
-    /// 配置 channel 转发器，用于把原始负载发往外部任务。
-    pub fn payload_sender(mut self, payload_tx: mpsc::UnboundedSender<Vec<u8>>) -> Self {
-        self.payload_tx = Some(payload_tx);
-        self
-    }
-
-    /// 注册原始事件处理器。
-    ///
-    /// - 传入 `"raw"` 会接收所有原始事件负载
-    /// - 传入具体 `event_type`（例如 `"im.message.receive_v1"`）只会接收匹配事件
-    pub fn register_raw<S, H>(mut self, key: S, handler: H) -> Result<Self, String>
-    where
-        S: Into<String>,
-        H: EventHandler,
-    {
-        let key = key.into();
-        if key.trim().is_empty() {
-            return Err("processor key cannot be empty".to_string());
-        }
-        if self.raw_handlers.contains_key(&key) {
-            return Err(format!("processor already registered, type: {key}"));
-        }
-        self.raw_handlers.insert(key, Arc::new(handler));
-        Ok(self)
-    }
-
-    fn extract_event_type(payload: &[u8]) -> Option<String> {
-        serde_json::from_slice::<RawEventEnvelope>(payload)
-            .ok()
-            .map(|event| event.header.event_type)
-            .filter(|event_type| !event_type.trim().is_empty())
-    }
-
-    fn dispatch_raw_handler(&self, key: &str, payload: &[u8]) -> Result<(), String> {
-        if let Some(handler) = self.raw_handlers.get(key) {
-            handler
-                .handle(payload)
-                .map_err(|err| format!("处理原始事件 {key} 失败: {err}"))?;
-        }
-        Ok(())
-    }
-
-    /// 在不做 schema 校验的前提下分发原始负载。
-    pub fn do_without_validation(&self, payload: &[u8]) -> Result<(), String> {
-        if let Some(payload_tx) = &self.payload_tx {
-            payload_tx
-                .send(payload.to_vec())
-                .map_err(|e| format!("转发事件负载失败: {e}"))?;
-        }
-
-        if let Some(event_type) = Self::extract_event_type(payload) {
-            self.dispatch_raw_handler(&event_type, payload)?;
-        }
-
-        self.dispatch_raw_handler(Self::RAW_EVENT_KEY, payload)?;
-
-        Ok(())
-    }
 }
 
 const END_POINT_URL: &str = "/callback/ws/endpoint";
@@ -286,124 +156,6 @@ pub(crate) struct EndPointResponse {
 pub(crate) struct ClientConfig {
     #[serde(rename = "PingInterval")]
     pub(crate) ping_interval: i32,
-}
-
-/// WebSocket 客户端结果类型别名。
-pub type WsClientResult<T> = Result<T, WsClientError>;
-
-#[derive(Debug, thiserror::Error)]
-/// WebSocket 客户端错误类型。
-pub enum WsClientError {
-    #[error("unexpected response")]
-    /// 返回体缺少预期字段。
-    UnexpectedResponse,
-    #[error("Request error: {0}")]
-    /// 端点查询 HTTP 请求失败。
-    RequestError(#[from] reqwest::Error),
-    #[error("Url parse error: {0}")]
-    /// WebSocket 地址解析失败。
-    UrlParseError(#[from] url::ParseError),
-    #[error("Server error: {code}, {message}")]
-    /// 服务端返回业务错误。
-    ServerError {
-        /// 服务端错误码。
-        code: i32,
-        /// 服务端错误描述。
-        message: String,
-    },
-    #[error("Client error: {code}, {message}")]
-    /// 客户端侧参数或状态错误。
-    ClientError {
-        /// 客户端错误码。
-        code: i32,
-        /// 客户端错误描述。
-        message: String,
-    },
-    #[error("connection closed")]
-    /// 连接被关闭。
-    ConnectionClosed {
-        /// 关闭原因。
-        reason: Option<WsCloseReason>,
-    },
-    #[error("WebSocket error: {0}")]
-    /// WebSocket 传输错误。
-    WsError(Box<tokio_tungstenite::tungstenite::Error>),
-    #[error("Prost error: {0}")]
-    /// Protobuf 解码错误。
-    ProstError(#[from] prost::DecodeError),
-    #[error("malformed control frame: {message}")]
-    /// 控制帧（如 pong）payload 非法。
-    MalformedControlFrame {
-        /// 错误描述。
-        message: String,
-    },
-    #[error("invalid session state transition: {kind}")]
-    /// 会话状态不允许的操作（例如已关闭后继续处理业务帧）。
-    InvalidStateTransition {
-        /// 结构化原因（测试与调用方可 match，勿依赖展示字符串）。
-        kind: InvalidStateKind,
-    },
-    #[error("event handler panicked")]
-    /// 用户 EventHandler 在 blocking 池中 panic。
-    HandlerPanicked,
-    #[error("handler backlog full: {message}")]
-    /// 串行 handler 队列与本地 outbox 均已满。
-    BacklogFull {
-        /// 错误描述。
-        message: String,
-    },
-    #[error("invalid frame method: {method}")]
-    /// 未知 protobuf frame method。
-    InvalidFrameMethod {
-        /// 收到的 method 值。
-        method: i32,
-    },
-}
-
-/// 会话非法状态转换的结构化原因。
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum InvalidStateKind {
-    /// 需要 Active，实际为其它状态。
-    ExpectedActive {
-        /// 当前状态名（如 `"Closing"` / `"Closed"`）。
-        actual: &'static str,
-    },
-    /// Closing 期间又收到业务 Binary/Text。
-    DataWhileClosing,
-    /// 已 Closed 仍尝试 begin_close。
-    AlreadyClosed,
-    /// 串行 handler worker 通道已关闭。
-    WorkerGone,
-}
-
-impl std::fmt::Display for InvalidStateKind {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::ExpectedActive { actual } => {
-                write!(f, "session is {actual}, expected Active")
-            }
-            Self::DataWhileClosing => {
-                write!(f, "received data frame while session is Closing")
-            }
-            Self::AlreadyClosed => write!(f, "session already Closed"),
-            Self::WorkerGone => write!(f, "handler worker is gone"),
-        }
-    }
-}
-
-impl From<tokio_tungstenite::tungstenite::Error> for WsClientError {
-    fn from(error: tokio_tungstenite::tungstenite::Error) -> Self {
-        WsClientError::WsError(Box::new(error))
-    }
-}
-
-/// 连接关闭原因。
-#[derive(Debug, Clone)]
-pub struct WsCloseReason {
-    /// WebSocket 关闭码。
-    pub code: CloseCode,
-    /// 关闭原因文案。
-    pub message: String,
 }
 
 #[cfg(test)]
