+++ crates/openlark-client/src/ws_client/mod.rs
@@ -4,6 +4,7 @@
 // 会话协议在单一 `session` loop 中实现；frame / package 为内部细节。
 
 mod client;
+mod dispatcher;
 mod frame_handler;
 mod headers;
 mod package;
@@ -11,10 +12,11 @@ mod session;
 
 #[cfg(feature = "websocket")]
 /// 会话级公开 API。
-pub use client::{
-    EventDispatcherHandler, EventHandler, InvalidStateKind, LarkWsClient, WsClientError,
-    WsClientResult, WsCloseReason,
-};
+pub use client::LarkWsClient;
+#[cfg(feature = "websocket")]
+pub use dispatcher::{EventDispatcherHandler, EventHandler};
+#[cfg(feature = "websocket")]
+pub use session::{InvalidStateKind, WsClientError, WsClientResult, WsCloseReason};
 
 #[cfg(test)]
 mod tests;
